+++ packages/nimble-components/src/table-column/date-text/tests/table-column-date-text.spec.ts
@@ -78,12 +78,15 @@ describe('TableColumnDateText', () => {
             expect(column.checkValidity()).toBeTrue();
         });
 
-        // WebKit skipped, see https://github.com/ni/nimble/issues/1940
-        it('changing fieldName updates display #SkipWebkit', async () => {
+        it('changing fieldName updates display', async () => {
+            const fieldValue = new Date('Dec 10, 2012, 10:35:05 PM').valueOf();
+            const anotherFieldValue = new Date(
+                'Jan 20, 2018, 4:05:45 AM'
+            ).valueOf();
             await table.setData([
                 {
-                    field: new Date('Dec 10, 2012, 10:35:05 PM').valueOf(),
-                    anotherField: new Date('Jan 20, 2018, 4:05:45 AM').valueOf()
+                    field: fieldValue,
+                    anotherField: anotherFieldValue
                 }
             ]);
             await waitForUpdatesAsync();
@@ -91,19 +94,22 @@ describe('TableColumnDateText', () => {
             column.fieldName = 'anotherField';
             await waitForUpdatesAsync();
 
+            const expectedFormattedValue = pageObject.getDefaultFormattedCellText(
+                anotherFieldValue,
+                'en-US'
+            );
             expect(pageObject.getRenderedCellContent(0, 0)).toEqual(
-                'Jan 20, 2018, 4:05:45 AM'
+                expectedFormattedValue
             );
         });
 
-        // WebKit skipped, see https://github.com/ni/nimble/issues/1940
-        it('changing data from value to null displays blank #SkipWebkit', async () => {
-            await table.setData([
-                { field: new Date('Dec 10, 2012, 10:35:05 PM').valueOf() }
-            ]);
+        it('changing data from value to null displays blank', async () => {
+            const fieldValue = new Date('Dec 10, 2012, 10:35:05 PM').valueOf();
+            await table.setData([{ field: fieldValue }]);
             await waitForUpdatesAsync();
+            const expectedFormattedValue = pageObject.getDefaultFormattedCellText(fieldValue, 'en-US');
             expect(pageObject.getRenderedCellContent(0, 0)).toEqual(
-                'Dec 10, 2012, 10:35:05 PM'
+                expectedFormattedValue
             );
 
             const updatedValue = { field: null };
@@ -114,19 +120,18 @@ describe('TableColumnDateText', () => {
             expect(pageObject.getRenderedCellContent(0, 0)).toEqual('');
         });
 
-        // WebKit skipped, see https://github.com/ni/nimble/issues/1940
-        it('changing data from null to value displays value #SkipWebkit', async () => {
+        it('changing data from null to value displays value', async () => {
             await table.setData([{ field: null }]);
             await waitForUpdatesAsync();
             expect(pageObject.getRenderedCellContent(0, 0)).toEqual('');
 
-            await table.setData([
-                { field: new Date('Dec 10, 2012, 10:35:05 PM').valueOf() }
-            ]);
+            const fieldValue = new Date('Dec 10, 2012, 10:35:05 PM').valueOf();
+            await table.setData([{ field: fieldValue }]);
             await waitForUpdatesAsync();
 
+            const expectedFormattedValue = pageObject.getDefaultFormattedCellText(fieldValue, 'en-US');
             expect(pageObject.getRenderedCellContent(0, 0)).toEqual(
-                'Dec 10, 2012, 10:35:05 PM'
+                expectedFormattedValue
             );
         });
 
@@ -140,12 +145,10 @@ describe('TableColumnDateText', () => {
             expect(pageObject.getRenderedCellContent(0, 0)).toEqual('');
         });
 
-        // WebKit skipped, see https://github.com/ni/nimble/issues/1940
-        it('sets title when cell text is ellipsized #SkipWebkit', async () => {
+        it('sets title when cell text is ellipsized', async () => {
             table.style.width = '200px';
-            await table.setData([
-                { field: new Date('Dec 10, 2012, 10:35:05 PM').valueOf() }
-            ]);
+            const fieldValue = new Date('Dec 10, 2012, 10:35:05 PM').valueOf();
+            await table.setData([{ field: fieldValue }]);
             await waitForUpdatesAsync();
             tablePageObject.dispatchEventToCell(
                 0,
@@ -153,8 +156,9 @@ describe('TableColumnDateText', () => {
                 new MouseEvent('mouseover')
             );
             await waitForUpdatesAsync();
+            const expectedFormattedValue = pageObject.getDefaultFormattedCellText(fieldValue, 'en-US');
             expect(pageObject.getCellTitle(0, 0)).toEqual(
-                'Dec 10, 2012, 10:35:05 PM'
+                expectedFormattedValue
             );
         });
 
@@ -193,14 +197,13 @@ describe('TableColumnDateText', () => {
             expect(pageObject.getCellTitle(0, 0)).toEqual('');
         });
 
-        // WebKit skipped, see https://github.com/ni/nimble/issues/1940
-        it('sets group header text to rendered date value #SkipWebkit', async () => {
-            await table.setData([
-                { field: new Date('Dec 10, 2012, 10:35:05 PM').valueOf() }
-            ]);
+        it('sets group header text to rendered date value', async () => {
+            const fieldValue = new Date('Dec 10, 2012, 10:35:05 PM').valueOf();
+            await table.setData([{ field: fieldValue }]);
             await waitForUpdatesAsync();
+            const expectedFormattedValue = pageObject.getDefaultFormattedCellText(fieldValue, 'en-US');
             expect(pageObject.getRenderedGroupHeaderContent(0)).toBe(
-                'Dec 10, 2012, 10:35:05 PM'
+                expectedFormattedValue
             );
         });
 
@@ -214,19 +217,19 @@ describe('TableColumnDateText', () => {
             expect(pageObject.getRenderedCellContent(0, 0)).toBe('12/10/2012');
         });
 
-        // WebKit skipped, see https://github.com/ni/nimble/issues/1940
-        it('updates displayed date when lang token changes #SkipWebkit', async () => {
-            await table.setData([
-                { field: new Date('Dec 10, 2012, 10:35:05 PM').valueOf() }
-            ]);
+        it('updates displayed date when lang token changes', async () => {
+            const fieldValue = new Date('Dec 10, 2012, 10:35:05 PM').valueOf();
+            await table.setData([{ field: fieldValue }]);
             await waitForUpdatesAsync();
+            const expectedEnglishFormattedValue = pageObject.getDefaultFormattedCellText(fieldValue, 'en-US');
             expect(pageObject.getRenderedCellContent(0, 0)).toBe(
-                'Dec 10, 2012, 10:35:05 PM'
+                expectedEnglishFormattedValue
             );
             lang.setValueFor(table, 'fr');
             await waitForUpdatesAsync();
+            const expectedFrenchFormattedValue = pageObject.getDefaultFormattedCellText(fieldValue, 'fr');
             expect(pageObject.getRenderedCellContent(0, 0)).toBe(
-                '10 déc. 2012, 22:35:05'
+                expectedFrenchFormattedValue
             );
         });